Suite de plancton et courants de surface
2024.07.10
•	Création de C_LandscapeMarine avec translate et checkBordure pour
éviter le ralentissement de particule
•	Passage de C_Plankton en A_Organisme plutôt que A_Animal
•	Procédure de sauvegarde d’écran
CaptureEcranPeriodique.captureEcranPlankton
2024.07.05
•	Ajout d’un calendrier
•	Lecture des courants
•	Ajout de C_MarineCell dans le contexte (elles activent le déplacement
du plancton
•	Modification de C_Landscape.translate pour prendre en compte les
organismes (plancton)
•	1ère simulation de particules :)
2024.07.04
•	Création de C_MarineCell avec champs courants nord et est
•	Création d’un readRaster double
•	Modification de l’arborescence raster pour les courants

+++ MelaneSim/src/data/converters/C_ConvertTimeAndSpace.java
@@ -65,7 +65,7 @@ public class C_ConvertTimeAndSpace {
 	}
 
 	/** Convert speed ex : 60km/h to X u_meter/tick */
-	public static double convertSpeed_UspaceByTick(int speedValue, String spaceUnit, String timeUnit) { // 60, km, h -> 60km/h
+	public static double convertSpeed_UspaceByTick(double speedValue, String spaceUnit, String timeUnit) { // 60, km, h -> 60km/h
 		return Math.round((speedValue * convertSpaceUnitToRasterUnit(spaceUnit) * (double) tick_Ucalendar)
 				/ _UcalendarTo_Utick(timeUnit));
 	}
